requirements.txt
PySimpleGUI==4.60.3
firebase-admin==5.3.0
opencv-python-headless==4.5.5.64
face_recognition==1.3.0
Pillow==8.4.0
numpy==1.21.2
cvzone==1.5.6